# PERF: Avoid materializing values in Categorical.set_categories