@TomAugspurger 2017-09-13
Here's the ASV:

```
[100.00%] ··· Running categoricals.Categoricals2.time_set_categories                                                                                80.8±1ms       before           after         ratio
     [f11bbf2f]       [6d72836e]
-        80.8±1ms       15.4±0.3ms     0.19  categoricals.Categoricals2.time_set_categories

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
```