@pep8speaks 2017-09-13
Hello @TomAugspurger! Thanks for updating the PR.





 - In the file [`pandas/tests/test_categorical.py`](https://github.com/pandas-dev/pandas/blob/f74da7e3d5d9a32efccd34905e36a9d5b3d9ab70/pandas/tests/test_categorical.py), following are the PEP8 issues :

> [Line 999:5](https://github.com/pandas-dev/pandas/blob/f74da7e3d5d9a32efccd34905e36a9d5b3d9ab70/pandas/tests/test_categorical.py#L999): [E303](https://duckduckgo.com/?q=pep8%20E303) too many blank lines (2)