@TomAugspurger 2017-09-14
I don't think so. `union_categoricals` takes the union of the two
categories. This replaces the old categories with the new.

On Wed, Sep 13, 2017 at 6:24 PM, Jeff Reback <notifications@github.com>
wrote:

> *@jreback* commented on this pull request.
> ------------------------------
>
> In pandas/core/categorical.py
> <https://github.com/pandas-dev/pandas/pull/17515#discussion_r138764715>:
>
> > +def _recode_for_categories(codes, old_categories, new_categories):
> +    """
> +    Convert a set of codes for to a new set of categories
> +
> +    Parameters
> +    ----------
> +    codes : array
> +    old_categories, new_categories : Index
> +
> +    Returns
> +    -------
> +    new_codes : array
> +
> +    Examples
> +    --------
> +    >>> old_cat = pd.Index(['b', 'a', 'c'])
>
> isn't this just a special case of union_categoricals?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/pandas-dev/pandas/pull/17515#pullrequestreview-62604925>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABQHIr3MDXGGPEIv_2U2nqZ4qxzcvvvRks5siGQ9gaJpZM4PWjZF>
> .
>